# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem manajemen Wira Teknik. Mendukung autentikasi, manajemen pengguna, dan fitur bisnis lainnya.",
        "title": "Wira Teknik API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Tim Wira Teknik",
            "email": "admin@wira-teknik.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/attachments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengunggah file lampiran (image/video/pdf) yang terkait dengan resource tertentu. Gunakan form-data dengan field: file, related_id, category. Kategori yang tersedia: shipment_delivery, shipment_received, invoice, payment_proof, bon, surat_jalan.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload lampiran",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File yang akan diunggah",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID resource terkait",
                        "name": "related_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "shipment_delivery",
                            "shipment_received",
                            "invoice",
                            "payment_proof",
                            "bon",
                            "surat_jalan"
                        ],
                        "type": "string",
                        "description": "Kategori lampiran",
                        "name": "category",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Attachment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/attachments/{relatedId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar lampiran (foto/video/pdf) yang terkait dengan resource tertentu (shipment, invoice, payment, dll)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attachments"
                ],
                "summary": "Ambil semua lampiran berdasarkan related ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Related ID (Shipment ID, Invoice ID, dll)",
                        "name": "relatedId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Attachment"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar semua invoice. Invoice otomatis dibuat saat pengiriman dibuat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Ambil semua invoice",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Invoice"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail invoice berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Ambil detail invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/shipments/{shipmentId}/invoice": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil invoice yang terkait dengan pengiriman tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Ambil invoice berdasarkan Shipment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "shipmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Attachment": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.AttachmentCategory"
                },
                "created_at": {
                    "type": "string"
                },
                "file_type": {
                    "$ref": "#/definitions/models.FileType"
                },
                "file_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "related_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.AttachmentCategory": {
            "type": "string",
            "enum": [
                "shipment_delivery",
                "shipment_received",
                "invoice",
                "payment_proof",
                "bon",
                "surat_jalan"
            ],
            "x-enum-varnames": [
                "AttachmentCategoryShipmentDelivery",
                "AttachmentCategoryShipmentReceived",
                "AttachmentCategoryInvoice",
                "AttachmentCategoryPaymentProof",
                "AttachmentCategoryBon",
                "AttachmentCategorySuratJalan"
            ]
        },
...
        "models.FileType": {
            "type": "string",
            "enum": [
                "image",
                "video",
                "pdf"
            ],
            "x-enum-varnames": [
                "FileTypeImage",
                "FileTypeVideo",
                "FileTypePdf"
            ]
        },
        "models.Invoice": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_no": {
                    "type": "string"
                },
                "payment_status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                },
                "remaining_balance": {
                    "type": "number"
                },
                "shipment_id": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "unpaid",
                "partial",
                "paid"
            ],
            "x-enum-varnames": [
                "PaymentStatusUnpaid",
                "PaymentStatusPartial",
                "PaymentStatusPaid"
            ]
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operasi berhasil"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}